% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dist.parallel}
\alias{dist.parallel}
\title{Parallel distance matrix}
\usage{
dist.parallel(tsList, distFunc = tsdiss.euclidean, cores = 2)
}
\arguments{
\item{tsList}{indepent variable with dimension T*1}

\item{distFunc}{dependent variable with dimension T*1}

\item{cores}{bandwidt}
}
\description{
Parallel distance matrix
}
\keyword{Parallel}
\keyword{distance}
\keyword{matrix}
